% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack.R
\name{blackjack}
\alias{blackjack}
\title{blackjack}
\usage{
blackjack()
}
\description{
A game of blackjack!
}
\details{
\describe{
\item{- minimum bet is 5}{}
\item{- dealer stands on 17 or higher}{}
\item{- no soft hands for dealer - Ace counts as 11}{}
\item{- one deck, shuffled once down to 20 cards or fewer}{}
\item{- no insurance}{}
\item{- no splits}{}
\item{- no additional payouts for natural blackjack}{}
}
}
