% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameRoom-package.R
\docType{package}
\name{gameRoom-package}
\alias{gameRoom}
\alias{gameRoom-package}
\title{gameRoom: Selection of Turn-Based Games}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tic Tac Toe, Hangman, Blackjack in the R Command Line.
}
\author{
\strong{Maintainer}: Weylandt Max \email{mweylandt@gmail.com}

}
\keyword{internal}
