% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play Games}
\usage{
play(game = NULL)
}
\arguments{
\item{game}{string input. One of 'tictactoe', 'hangman', or 'blackjack'}
}
\description{
This function calls the code required to run the available games
}
\examples{
\dontrun{play('tictactoe')}
}
