% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack.R
\name{blackjack}
\alias{blackjack}
\title{blackjack}
\usage{
blackjack()
}
\description{
A game of blackjack!
}
\details{
\describe{
\item{- minimum bet is 5}{}
\item{- no splits}{}
\item{- one deck, shuffled once down to 20 cards or fewer}{}


}
}
